move: 주스탠드 스토어 폴더이동
